feat: fcm 토큰 전달 카드 등록할 때로 변경 (#74)
