+++ lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:dio/dio.dart';
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:get/get.dart';
@@ -20,25 +19,9 @@ import 'package:once_front/src/screens/mypage/register_maincard.dart';
 import 'package:once_front/src/screens/mypage/user_edit.dart';
 import 'package:once_front/src/screens/mywallet/mywallet.dart';
 
-import 'package:firebase_messaging/firebase_messaging.dart';
-
 import 'constants.dart';
 
 void main() async {
-  print("1");
-  WidgetsFlutterBinding.ensureInitialized();
-  print("2");
-  await Firebase.initializeApp(); // Firebase 초기화
-  print("3");
-  final token = await FirebaseMessaging.instance.getToken();
-  print("4");
-  if (token != null) {
-    print("FCM Token: $token");
-    await sendTokenToServer(token); // 서버에 토큰 전송
-  } else {
-    print("FCM Token이 null입니다.");
-  }
-
   runApp(const MyApp());
   getPerformanceData();
 }
@@ -100,32 +83,5 @@ class MyApp extends StatelessWidget {
         GetPage(name: "/signup/4", page: () => Signup4()),
       ],
     );
-  }
-}
-
-Future<void> sendTokenToServer(String token) async {
-  final String BASE_URL = Constants.baseUrl;
-  final String apiUrl = '$BASE_URL/user/token';
-
-  const storage = FlutterSecureStorage();
-  String? storedAccessToken = await storage.read(key: 'accessToken');
-
-  if (storedAccessToken != null) {
-    final baseOptions = BaseOptions(
-      headers: {'Authorization': 'Bearer $storedAccessToken'},
-    );
-
-    final dio = Dio(baseOptions);
-
-    try {
-      var response =
-      await dio.post(apiUrl, data: {"token": token});
-      Map<dynamic, dynamic> responseData = response.data;
-      print(responseData);
-    } catch (e) {
-      print('토큰 전송 오류: $e');
-    }
-  } else {
-    print('AccessToken 없음');
   }
 }
