+++ pubspec.yaml
@@ -32,10 +32,6 @@ dependencies:
   flutter_secure_storage: ^4.2.0
   flutter_spinkit: ^5.2.1
   flutter_inappwebview: ^5.3.2
-  firebase_core: ^2.5.0
-  firebase_messaging: ^14.0.0
-  firebase_auth: ^4.12.0
-
 
 dev_dependencies:
   flutter_test:
